refactor: decouple middlewares from server
