+++ app/server.go
@@ -1,18 +1,17 @@
 package app
 
 import (
-	"context"
 	"encoding/json"
 	"net/http"
 
-	"github.com/adfinis-sygroup/mopsos/app/models"
-	otelObs "github.com/cloudevents/sdk-go/observability/opentelemetry/v2/client"
-	"github.com/cloudevents/sdk-go/v2/binding"
 	"github.com/cloudevents/sdk-go/v2/event"
-	httproto "github.com/cloudevents/sdk-go/v2/protocol/http"
 	http_logrus "github.com/improbable-eng/go-httpwares/logging/logrus"
 	"github.com/sirupsen/logrus"
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
+
+	"github.com/adfinis-sygroup/mopsos/app/middleware"
+	"github.com/adfinis-sygroup/mopsos/app/models"
+	"github.com/adfinis-sygroup/mopsos/app/types"
 )
 
 // Server is the main webserver struct
@@ -21,12 +20,6 @@ type Server struct {
 
 	EventChan chan<- models.EventData
 }
-
-type eventContext string
-
-var ContextUsername eventContext = "mopsos.username"
-var ContextEvent eventContext = "mopsos.event"
-var ContextRecord eventContext = "mopsos.record"
 
 // NewServer creates a server that receives CloudEvents from the network
 func NewServer(cfg *Config) *Server {
@@ -40,12 +33,14 @@ func (s *Server) Start() {
 	mux := http.NewServeMux()
 	mux.HandleFunc("/health", s.HandleHealthCheck)
 	mux.Handle("/webhook", otelhttp.NewHandler(
-		s.Authenticate(
-			s.LoadEvent(
-				s.Validate(
+		middleware.Authenticate(
+			middleware.LoadEvent(
+				middleware.Validate(
 					http.HandlerFunc(s.HandleWebhook),
 				),
+				s.config.EnableTracing,
 			),
+			s.config.BasicAuthUsers,
 		),
 		"webhook-receiver"),
 	)
@@ -63,73 +58,6 @@ func (s *Server) WithEventChannel(eventChan chan<- models.EventData) *Server {
 	return s
 }
 
-// Authenticate middleware handles checking credentials
-func (s *Server) Authenticate(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// get basic auth credentials
-		username, password, ok := r.BasicAuth()
-		if !ok {
-			http.Error(w, "missing Authorization header", http.StatusUnauthorized)
-			return
-		}
-
-		logrus.WithFields(logrus.Fields{
-			"username": username,
-		}).Debug("checking credentials")
-		if s.config.BasicAuthUsers[username] != password {
-			http.Error(w, "invalid credentials", http.StatusUnauthorized)
-			return
-		}
-		ctx := context.WithValue(r.Context(), ContextUsername, username)
-
-		next.ServeHTTP(w, r.WithContext(ctx))
-	})
-}
-
-// LoadEvent middlerware loads event from the request
-func (s *Server) LoadEvent(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// get event
-		message := httproto.NewMessageFromHttpRequest(r)
-		event, err := binding.ToEvent(r.Context(), message)
-		if err != nil {
-			logrus.WithError(err).Error("failed to decode event")
-			return
-		}
-		if s.config.EnableTracing {
-			// inject the span context into the event so it can be use i.e. while inserting to the database
-			otelObs.InjectDistributedTracingExtension(r.Context(), *event)
-		}
-		logrus.Debugf("received event: %v", event)
-
-		ctx := context.WithValue(r.Context(), ContextEvent, event)
-		next.ServeHTTP(w, r.WithContext(ctx))
-	})
-}
-
-// Validate middleware handles checking received events for validity
-func (s *Server) Validate(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		event := r.Context().Value(ContextEvent).(*event.Event)
-		record := &models.Record{}
-
-		if err := event.DataAs(record); err != nil {
-			logrus.WithError(err).Errorf("failed to unmarshal event data")
-			http.Error(w, "failed to unmarshal event data", http.StatusInternalServerError)
-			return
-		}
-
-		// reject record that have not been sent from the right auth
-		if record.ClusterName != r.Context().Value(ContextUsername).(string) {
-			http.Error(w, "event data does not match username", http.StatusUnauthorized)
-			return
-		}
-
-		ctx := context.WithValue(r.Context(), ContextRecord, record)
-		next.ServeHTTP(w, r.WithContext(ctx))
-	})
-}
-
 func (s *Server) HandleHealthCheck(w http.ResponseWriter, r *http.Request) {
 	// an example API handler
 	err := json.NewEncoder(w).Encode(map[string]bool{"ok": true})
@@ -140,8 +68,8 @@ func (s *Server) HandleHealthCheck(w http.ResponseWriter, r *http.Request) {
 
 func (s *Server) HandleWebhook(w http.ResponseWriter, r *http.Request) {
 	// get middleware data from context
-	event := r.Context().Value(ContextEvent).(*event.Event)
-	record := r.Context().Value(ContextRecord).(*models.Record)
+	event := r.Context().Value(types.ContextEvent).(*event.Event)
+	record := r.Context().Value(types.ContextRecord).(*models.Record)
 
 	// send the event to the main app via the async channel
 	s.EventChan <- models.EventData{
