+++ app/types/eventContext.go
@@ -1,0 +1,7 @@
+package types
+
+type eventContext string
+
+var ContextUsername eventContext = "mopsos.username"
+var ContextEvent eventContext = "mopsos.event"
+var ContextRecord eventContext = "mopsos.record"
